pkg/prometheus/ha: split up sharding code into its own file
